Update test.lisp
Adjusted according to the need for the new 3rd arg to SAFE-READ being NIL, stating non-interactive mode. Doing so, bypasses the use of cached buffers entirely, and avoids the capture and EQL comparison of stream args that may have been DX.

When expressions are complete in a single line of text, the non-interactive mode will work as expected. But if you split the text across multiple lines, as in test-5, then it will not work. This is an artifact of the line-oriented input in the initial call to READ-LIMITED-LINE.

There is a bit of a mismatch between Lisp READ and your SAFE-READ in this non-interactive mode. Lisp READ is not line oriented and will happily scan through newline characters embedded in a string during READ-FROM-STRING. Lisp READ is just looking for matching parens and ignores newlines as whitespace chars. So by forcing SAFE-READ to perform one line at a time, you are working at cross purposes against the Lisp READ.

As I understand your algorithm now, I see that you are attempting to limit the length of input by using your READ-LIMITED-LINE. 

But when dealing expressly with string input, you could bypass the line-oriented checking altogether. Just check the length of the input string, and forego conversion to stream from WITH-INPUT-FROM-STRING, calling on Lisp READ-FROM-STRING directly, under the control of your limited readtable. This is what I actually needed, and so rather than chopping up your code any further, I just copied the vital pieces from you, and pasted them into a separate lib for my purposes.