Allow input stream arg to be a string
Avoids conversion of strings to stream, avoids dangling DX references altogether, and allows for arbitrary multi-line string input all at once.

When users already have a string to be scanned, they no longer have to convert to a stream via WITH-INPUT-FROM-STRING, and can just call SAFE-READ directly with their string argument. But if they *do* have an idea to use WITH-INPUT-FROM-STRING, they should be sternly warned that this is an unsafe practice.