3rd Proposed Fix...
Have SAFE-READ return a third value on :INCOMPLETE-INPUT, being the buffered input needed for a repeat call to SAFE-READ. 

Have SAFE-READ accept a third arg, BUFFER, defaulting to NIL, and should be the returned BUFFER third arg from the previous call to SAFE-READ, when you intend to re-read additional input from the stream to fix the :INCOMPLETE-INPUT condition.

Normal use of SAFE-READ simply calls with none, one, or two args. Only rectifying repeat-reads need to supply the third arg.

In this way, we make the caller responsible for the scope and lifetime of the stream. We do not cache anything locally, and hence no unsafe practices no matter what stream they throw at us.

(I would still lobby for a second STRING, in addition to STREAM, entry point, to forego the need for line-oriented input when you already know the future.)